set the alarm window to 2000 on startup
git-svn-id: svn://orca.physics.unc.edu/Orca@11627 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Objects/Hardware/Serial/RefClock/ORSynClockModel.h
@@ -57,7 +57,7 @@
 - (int) syncMode;
 - (void) setSyncMode:(int)aMode;
 - (unsigned long) alarmWindow;
-- (void) setAlarmWindow:(unsigned int)aValue;
+- (void) setAlarmWindow:(unsigned long)aValue;
 
 #pragma mark ***Commands
 - (void) writeData:(NSDictionary*)aDictionary;
